clear index.d.ts add true declarations
