+++ types/muvjs/tsconfig.json
@@ -4,6 +4,8 @@
     "sourceMap": true,
     "noImplicitAny": true,
     "strictNullChecks": true,
+    "noImplicitThis": true,
+    "strictFunctionTypes": true,
     "alwaysStrict": true,
     "module": "es6",
     "target": "es5",
@@ -11,8 +13,6 @@
       "es2019",
       "dom"
     ],
-    "noEmit": false,
-    "declaration": true,
     "allowJs": true,
     "checkJs": true,
     "baseUrl": "types",
